+++ src/PlcSimulation.cs
@@ -81,8 +81,8 @@ namespace OpcPlc
                 _boiler1Generator = _plcServer.TimeService.NewTimer(_plcServer.PlcNodeManager.UpdateBoiler1, 1000);
             }
 
-            // Start simulation of nodes from node list.
-            foreach (var nodes in NodesList)
+            // Start simulation of nodes from plugin node list.
+            foreach (var nodes in Program.PluginNodes)
             {
                 nodes.StartSimulation();
             }
@@ -98,8 +98,8 @@ namespace OpcPlc
             Disable(_eventInstanceGenerator);
             Disable(_boiler1Generator);
 
-            // Stop simulation of nodes from node list.
-            foreach (var nodes in NodesList)
+            // Stop simulation of nodes from plugin node list.
+            foreach (var nodes in Program.PluginNodes)
             {
                 nodes.StopSimulation();
             }
