+++ src/PluginNodes/LongIdNode.cs
@@ -1,13 +1,13 @@
-﻿namespace OpcPlc.Nodes
+﻿namespace OpcPlc.PluginNodes
 {
     using Opc.Ua;
     using System.Collections.Generic;
-    using System.Web;
+    using System.Text;
 
     /// <summary>
-    /// Node with special chars in name and ID.
+    /// Node with ID of 3950 chars.
     /// </summary>
-    public class SpecialCharNameNodes : INodes
+    public class LongIdNode : IPluginNodes
     {
         public IReadOnlyCollection<string> NodeIDs { get; private set; } = new List<string>();
 
@@ -18,8 +18,8 @@
         public void AddOption(Mono.Options.OptionSet optionSet)
         {
             optionSet.Add(
-                "scn|specialcharname",
-                $"add node with special characters in name.\nDefault: {_isEnabled}",
+                "lid|longid",
+                $"add node with ID of 3950 chars.\nDefault: {_isEnabled}",
                 (string p) => _isEnabled = p != null);
         }
 
@@ -51,13 +51,18 @@
 
         private void AddNodes(FolderState folder)
         {
-            string SpecialChars = HttpUtility.HtmlDecode(@"&quot;!&#167;$%&amp;/()=?`&#180;\+~*&#39;#_-:.;,&lt;&gt;|@^&#176;€&#181;{[]}");
+            // Repeat A-Z until 3950 chars are collected.
+            var id = new StringBuilder(4000);
+            for (int i = 0; i < 3950; i++)
+            {
+                id.Append((char)(65 + (i % 26)));
+            }
 
             _node = _plcNodeManager.CreateVariableNode<uint>(
                 _plcNodeManager.CreateBaseVariable(
                     folder,
-                    path: "Special_" + SpecialChars,
-                    name: SpecialChars,
+                    path: id.ToString(),
+                    name: "LongId3950",
                     new NodeId((uint)BuiltInType.UInt32),
                     ValueRanks.Scalar,
                     AccessLevels.CurrentReadOrWrite,
@@ -67,7 +72,7 @@
 
             NodeIDs = new List<string>
             {
-                "Special_" + SpecialChars,
+                id.ToString(),
             };
         }
     }
