+++ src/PluginNodes/SpecialCharNameNode.cs
@@ -1,13 +1,13 @@
-﻿namespace OpcPlc.Nodes
+﻿namespace OpcPlc.PluginNodes
 {
     using Opc.Ua;
     using System.Collections.Generic;
-    using System.Text;
+    using System.Web;
 
     /// <summary>
-    /// Node with ID of 3950 chars.
+    /// Node with special chars in name and ID.
     /// </summary>
-    public class LongIdNodes : INodes
+    public class SpecialCharNameNode : IPluginNodes
     {
         public IReadOnlyCollection<string> NodeIDs { get; private set; } = new List<string>();
 
@@ -18,8 +18,8 @@
         public void AddOption(Mono.Options.OptionSet optionSet)
         {
             optionSet.Add(
-                "lid|longid",
-                $"add node with ID of 3950 chars.\nDefault: {_isEnabled}",
+                "scn|specialcharname",
+                $"add node with special characters in name.\nDefault: {_isEnabled}",
                 (string p) => _isEnabled = p != null);
         }
 
@@ -51,18 +51,13 @@
 
         private void AddNodes(FolderState folder)
         {
-            // Repeat A-Z until 3950 chars are collected.
-            var id = new StringBuilder(4000);
-            for (int i = 0; i < 3950; i++)
-            {
-                id.Append((char)(65 + (i % 26)));
-            }
+            string SpecialChars = HttpUtility.HtmlDecode(@"&quot;!&#167;$%&amp;/()=?`&#180;\+~*&#39;#_-:.;,&lt;&gt;|@^&#176;€&#181;{[]}");
 
             _node = _plcNodeManager.CreateVariableNode<uint>(
                 _plcNodeManager.CreateBaseVariable(
                     folder,
-                    path: id.ToString(),
-                    name: "LongId3950",
+                    path: "Special_" + SpecialChars,
+                    name: SpecialChars,
                     new NodeId((uint)BuiltInType.UInt32),
                     ValueRanks.Scalar,
                     AccessLevels.CurrentReadOrWrite,
@@ -72,7 +67,7 @@
 
             NodeIDs = new List<string>
             {
-                id.ToString(),
+                "Special_" + SpecialChars,
             };
         }
     }
